Lindsey

+++ frontend/src/components/EventShow.vue
@@ -19,7 +19,7 @@
     <div class="show-right">
       <div class="show-description">{{ event.description }}</div>
       <ul class="event-requirement-list">
-          <li  class="event-requirement" v-for='requirement in requirements'>{{ requirement.title }}</li>
+          <li  class="event-requirement" v-for='requirement in requirement_titles'>{{ requirement }}</li>
       </ul>
       <div class="show-company">{{ showAdmin(event.admin_name,event.company_name) }}
         <div class=""><a :href="event.company_website">Company Website</a></div>
@@ -75,7 +75,7 @@ export default {
       applyModal.style.display = 'none'
     },
     submit () {
-      if (this.$store.state.request) {
+      if (this.$store.state.request && this.$store.state.request.event_id) {
         $.ajax({
           method: 'DELETE',
           url: '/api/requests/' + this.event.id,
@@ -104,7 +104,7 @@ export default {
       }
     },
     showButton () {
-      if (this.$store.state.request) {
+      if (this.$store.state.request && this.$store.state.request.event_id) {
         return ('Cancel Application')
       } else {
         return ('Apply')
@@ -115,8 +115,22 @@ export default {
     event: function () {
       return this.$store.state.event
     },
-    requirements: function () {
-      return this.event.requirements
+    requirement_titles: function () {
+      const titles = []
+      if (this.event.requirements) {
+        this.event.requirements.forEach(id => {
+          if (id === 1) {
+            titles.push('no visible tattoos')
+          } else if (id === 2) {
+            titles.push('valid driver\'s license')
+          } else if (id === 3) {
+            titles.push('valid TIPS certification')
+          } else if (id === 4) {
+            titles.push('no felonies')
+          }
+        })
+      }
+      return titles
     }
   }
 }
